責務に応じて関数やファイルを分割
