+++ lib/model/entity/Participant.dart
@@ -1,0 +1,5 @@
+class Participant {
+  String displayName;
+
+  Participant(this.displayName);
+}
